Refactor: Separated compression and database logics to their own packages.
